extract method: post_multipart_related
